Remove unnecessary casts
On all currently supported platforms c_float = f32, c_int = i32

+++ src/lib.rs
@@ -118,7 +118,7 @@ impl ImGui {
    }
    pub fn mouse_pos(&self) -> (f32, f32) {
       let io: &mut ffi::ImGuiIO = unsafe { mem::transmute(ffi::igGetIO()) };
-      (io.mouse_pos.x as f32, io.mouse_pos.y as f32)
+      (io.mouse_pos.x, io.mouse_pos.y)
    }
    pub fn set_mouse_pos(&mut self, x: f32, y: f32) {
       let io: &mut ffi::ImGuiIO = unsafe { mem::transmute(ffi::igGetIO()) };
@@ -129,8 +129,8 @@ impl ImGui {
       let io: &mut ffi::ImGuiIO = unsafe { mem::transmute(ffi::igGetIO()) };
       io.mouse_down = *states;
    }
-   pub fn get_time(&self) -> f32 { unsafe { ffi::igGetTime() as f32 } }
-   pub fn get_frame_count(&self) -> i32 { unsafe { ffi::igGetFrameCount() as i32 } }
+   pub fn get_time(&self) -> f32 { unsafe { ffi::igGetTime() } }
+   pub fn get_frame_count(&self) -> i32 { unsafe { ffi::igGetFrameCount() } }
    pub fn frame<'fr, 'a: 'fr>(&'a mut self, width: u32, height: u32, delta_time: f32) -> Frame<'fr> {
       unsafe {
          let io: &mut ffi::ImGuiIO = mem::transmute(ffi::igGetIO());
@@ -234,12 +234,12 @@ impl<'fr> Frame<'fr> {
    pub fn separator(&self) { unsafe { ffi::igSeparator() }; }
    pub fn same_line(&self, pos_x: f32) {
       unsafe {
-         ffi::igSameLine(pos_x as c_float, -1.0f32 as c_float)
+         ffi::igSameLine(pos_x, -1.0f32)
       }
    }
    pub fn same_line_spacing(&self, pos_x: f32, spacing_w: f32) {
       unsafe {
-         ffi::igSameLine(pos_x as c_float, spacing_w as c_float)
+         ffi::igSameLine(pos_x, spacing_w)
       }
    }
    pub fn spacing(&self) { unsafe { ffi::igSpacing() }; }
